change name of files, added the drag and drop feature(useing mouse up, down, move. Also moved everything related to object out of grid and into ObjectTokenFinal.js
