+++ team/m3/js/grid.js
@@ -1,44 +1,3 @@
-import {DataForm} from "../js/DataForm.js";
-import {DatabaseConnection } from "../js/DatabaseConnection.js";
-import {Image} from "../js/image.js";
-
-const battleGrid = document.getElementById('battle-grid');
-//start of emily's edit
-//initializing indexdb for object and image, initializing the object token and image import class
-const objectDB = new DatabaseConnection("ObjectStore");
-objectDB.openDatabase();
-const imageDB = new DatabaseConnection("ImageStore");
-
-const ObjForm = document.getElementById("object_form");
-ObjForm.style.display= "none";
-
-const dataObjForm = new DataForm(objectDB);
-dataObjForm.clickForm()
-
-const img = new Image(document.getElementById("object_form"), imageDB)
-const imgInput = document.getElementById("imageInput");
-imgInput.addEventListener("change", (event) => img.PreviewImg(event))
-
-const objectGrid = document.getElementById('object-grid');
-objectGrid.style.zIndex="1"
-
-//Switch between battle and object grid
-const switchButton = document.getElementById("switch_button");
-switchButton.addEventListener("click", function x(){
-    if(objectGrid.style.zIndex==="1"){
-        objectGrid.style.zIndex = "0";
-        battleGrid.style.zIndex = "1";
-        switchButton.value = "switch grid:Object-Grid"
-    }
-    else{
-        battleGrid.style.zIndex = "0";
-        objectGrid.style.zIndex = "1";
-        switchButton.value = "switch grid:Battle-Grid"
-    }
-})
-//end of emily's edit
-
-
 document.getElementById('create-grid').addEventListener('click', function() {
     const width = parseInt(document.getElementById('grid-width').value);
     const height = parseInt(document.getElementById('grid-height').value);
@@ -57,7 +16,6 @@ function createGrid(width, height) {
     objectGrid.innerHTML = ''; // Clear any existing grid
     objectGrid.style.gridTemplateColumns = `repeat(${width}, 1fr)`;
     objectGrid.style.gridTemplateRows = `repeat(${height}, 1fr)`;
-    const c = dataObjForm.addWH(width, height);
     //end of emily's edit
 
     for (let y = 0; y < height; y++) {
@@ -84,54 +42,4 @@ function createGrid(width, height) {
             battleGrid.appendChild(tile);
         }
     }
-
-    //start of emily's edit
-    //add the event listener for add/update object form, including delete, cancel, create and update
-    dataObjForm.render();
-    img.render();
-    dataObjForm.renderWhenLoad()
-    .then(result => {
-        result.id.forEach(i => console.log(i))
-    })
-
-    const createButton = document.getElementById("create");
-    const updateButton = document.getElementById("update");
-    const deleteButton = document.getElementById("delete");
-    const cancelButton = document.getElementById("cancel");
-
-    cancelButton.addEventListener("click", () => {
-        dataObjForm.clearForm();
-        img.clearImage();
-    })
-
-    createButton.addEventListener("click", (event) => {
-        const numCopy = document.getElementById("copy").value;
-        for(let i = 0; i<numCopy; i++){
-            dataObjForm.createObject(event)
-            .then(result => {
-                result.id.forEach(i => {
-                    img.createImageElement(document.getElementById(String(i)));
-                })
-            })
-        } 
-    });
-
-    updateButton.addEventListener("click", (event) => {
-        dataObjForm.updateObject()
-        .then(result => {
-            result.id.forEach(i => {
-                img.updateImageElement(document.getElementById(String(i)))        
-            })
-        })
-    });
-
-    deleteButton.addEventListener("click", (event) => {
-        dataObjForm.deleteObject()
-        .then(result => {
-            result.id.forEach(i => {
-                img.deleteImageElement(i)
-            })
-        })    
-    });
-    //end of emily's edit
 }
