+++ team/m3/main.html
@@ -88,5 +88,8 @@
 
     <script src="js/paint.js"></script>
 
+    <script type="module" src="js/ObjectTokenFinal.js"></script>
+
+
 </body>
 </html>
