+++ team/m3/styles.css
@@ -113,6 +113,7 @@ body {
     align-items: center;
     justify-content: center;
     cursor: pointer;
+    position: absolute;
   }
   
   
@@ -145,13 +146,15 @@ body {
     display: none;
   }
   
-  .tooltip {
+  .hoverBox {
     position: relative; 
     display: inline-block; 
     cursor: pointer; 
+    z-index: 4;
+
   }
   
-  .tooltip .tooltiptext {
+  .hoverBox .hovertext {
     visibility: hidden;
     position: absolute;
     padding: 5px;
@@ -169,9 +172,10 @@ body {
     transform: translateX(-50%);
   }
   
-  .tooltip:hover .tooltiptext {
+  .hoverBox:hover .hovertext {
     visibility: visible;
   }
+  
   
   #preview{
     object-fit: fill;
